requirements.txt
absl-py==2.3.0
ale-py==0.11.1
annotated-types==0.7.0
anyio==4.9.0
backtrader==1.9.78.123
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
contourpy==1.3.2
curl_cffi==0.11.4
cycler==0.12.1
Farama-Notifications==0.0.4
fastapi==0.115.14
filelock==3.18.0
Flask==3.1.1
fonttools==4.58.4
frozendict==2.4.6
fsspec==2025.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.0
gymnasium==1.1.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
multitasking==0.0.11
networkx==3.5
numpy==2.3.1
ollama==0.5.1
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.8
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pygame==2.6.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
rich==14.0.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stable_baselines3==2.6.0
starlette==0.46.2
sympy==1.14.0
ta-lib==0.6.4
tensorboard==2.19.0
tensorboard-data-server==0.7.2
torch==2.7.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.34.3
websockets==15.0.1
Werkzeug==3.1.3
yfinance==0.2.63
